move cross_entropy from fluid to phi
