+++ paddle/fluid/operators/math/CMakeLists.txt
@@ -21,7 +21,6 @@ else()
   math_library(concat_and_split DEPS concat_and_split_functor)
 endif()
 math_library(context_project DEPS im2col math_function)
-math_library(cross_entropy)
 math_library(cos_sim_functor)
 math_library(depthwise_conv)
 math_library(im2col)
