+++ paddle/phi/kernels/gpu/cross_entropy_kernel.cu
@@ -22,7 +22,6 @@ limitations under the License. */
 namespace cub = hipcub;
 #endif
 
-#include "paddle/fluid/operators/math/cross_entropy.h"
 #include "paddle/fluid/operators/math/softmax.h"
 #include "paddle/fluid/platform/device/gpu/gpu_dnn.h"
 #include "paddle/phi/backends/gpu/gpu_device_function.h"
@@ -31,6 +30,7 @@ namespace cub = hipcub;
 #include "paddle/phi/core/tensor_utils.h"
 #include "paddle/phi/core/visit_type.h"
 #include "paddle/phi/kernels/funcs/axis_utils.h"
+#include "paddle/phi/kernels/funcs/cross_entropy.h"
 #include "paddle/phi/kernels/funcs/for_range.h"
 #include "paddle/phi/kernels/funcs/math_function.h"
 #include "paddle/phi/kernels/gpudnn/softmax_gpudnn.h"
@@ -46,7 +46,7 @@ template <typename T>
 static __device__ __forceinline__ T Log(T x) {
   using AccT = typename dtype::MPTypeTrait<T>::Type;
   AccT logx = std::log(static_cast<AccT>(x));
-  return paddle::operators::math::TolerableValue<T>()(static_cast<T>(logx));
+  return phi::funcs::TolerableValue<T>()(static_cast<T>(logx));
 }
 
 // Wrapper of exp function. Use exp(float32) for float16
@@ -54,7 +54,7 @@ template <typename T>
 static __device__ __forceinline__ T Exp(T x) {
   using AccT = typename dtype::MPTypeTrait<T>::Type;
   AccT expx = std::exp(static_cast<AccT>(x));
-  return paddle::operators::math::TolerableValue<T>()(static_cast<T>(expx));
+  return phi::funcs::TolerableValue<T>()(static_cast<T>(expx));
 }
 
 template <typename Tx, typename Ty = Tx>
@@ -1285,16 +1285,15 @@ void CrossEntropyWithSoftmaxCUDAKernel(const GPUContext& dev_ctx,
     DenseTensor softmax_out_2d(*softmax_out);
     softmax_out_2d.Resize({n, d});
 
-    // math::CrossEntropyFunctor support axis is the last
+    // phi::funcs::CrossEntropyFunctor support axis is the last
     if (axis_v == -1) {
-      paddle::operators::math::CrossEntropyFunctor<GPUContext, T>()(
-          dev_ctx,
-          &loss_2d,
-          &softmax_2d,
-          &labels_2d,
-          soft_label,
-          ignore_index,
-          axis_dim);
+      phi::funcs::CrossEntropyFunctor<GPUContext, T>()(dev_ctx,
+                                                       &loss_2d,
+                                                       &softmax_2d,
+                                                       &labels_2d,
+                                                       soft_label,
+                                                       ignore_index,
+                                                       axis_dim);
       return;
     }
 
@@ -1389,14 +1388,13 @@ void CrossEntropyWithSoftmaxCUDAKernel(const GPUContext& dev_ctx,
       loss_2d.Resize({n, 1});
       paddle::operators::math::SoftmaxCUDNNFunctor<T, GPUContext>()(
           dev_ctx, &logits_2d, &softmax_2d);
-      paddle::operators::math::CrossEntropyFunctor<GPUContext, T>()(
-          dev_ctx,
-          &loss_2d,
-          &softmax_2d,
-          &labels_2d,
-          false,
-          ignore_index,
-          axis_dim);
+      phi::funcs::CrossEntropyFunctor<GPUContext, T>()(dev_ctx,
+                                                       &loss_2d,
+                                                       &softmax_2d,
+                                                       &labels_2d,
+                                                       false,
+                                                       ignore_index,
+                                                       axis_dim);
     } else {
       auto* logits_data = logits.data<T>();
       auto* labels_data = label.data<LabelT>();
